requirements.txt
Django==6.0.3
whitenoise==6.12.0
python-dotenv==1.2.2
